requirements.txt
# Framework Web
Flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
eventlet==0.35.2

# OpenStack clients
python-openstackclient==6.3.0
python-heatclient==3.3.0
python-novaclient==18.3.0
python-neutronclient==11.0.0
python-keystoneclient==5.1.0

# Base de données et configuration
python-dotenv==1.0.0

# Git pour import templates
GitPython==3.1.40

# YAML
PyYAML==6.0.1

# Validation
jsonschema==4.20.0

# Tests
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilitaires
requests==2.31.0
python-dateutil==2.8.2
werkzeug==3.0.1
